pyproject.toml
[project]
name = "vd-dbt-fabricspark"
dynamic = ["version"]
description = "A Microsoft Fabric Spark adapter plugin for dbt"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Pradeep Srikakolapu", email = "pvenkat@microsoft.com" }
]

keywords = [
    "dbt", "adapter", "adapters", "database", "elt", "dbt-core", "dbt Core", 
    "dbt Cloud", "dbt Labs", "spark", "microsoft", "fabric", "fabricspark", 
    "fabricspark adapter", "fabricspark dbt", "fabricspark dbt adapter", 
    "fabricspark dbt adapters", "fabricspark dbt core", "fabricspark dbt cloud", 
    "fabricspark dbt labs", "Fabric lakehouse", "lakehouse", "lakehouse dbt", 
    "lakehouse dbt adapter", "lakehouse dbt adapters", "lakehouse dbt core", 
    "lakehouse dbt cloud", "lakehouse dbt labs"
]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13"
]

requires-python = ">=3.9,<3.14"

dependencies = [
    "dbt-common>=1.10,<2.0",
    "dbt-adapters>=1.7,<2.0",
    # add dbt-core to ensure backwards compatibility of installation, this is not a functional dependency
    "dbt-core>=1.8.0",  
    "azure-identity>=1.13.0",
    "azure-core>=1.26.4",
    "azure-cli>=2.60.0",
    "PyJWT[crypto]>=2.8.0",
    "requests>=2.31.0",
]

[dependency-groups]
dev = [
    # Testing
    "pytest~=7.4",
    "pytest-dotenv~=0.5.2",
    "flaky~=3.7",
    "pytest-xdist~=3.5",
    #"dbt-tests-adapter @ git+https://github.com/dbt-labs/dbt-adapters.git@v1.10.0#egg=dbt-tests-adapter&subdirectory=dbt-tests-adapter",
    "dbt-tests-adapter",
    # Linting & formatting
    "ruff>=0.11.2",
    # Packaging
    "twine~=4.0",
    # Adapter specific dependencies
    "mock~=5.1",
]

[tool.uv]
default-groups = "all"
prerelease = "allow"

[project.urls]
homepage = "https://github.com/microsoft/dbt-fabricspark"
"Setup & configuration" = "https://docs.getdbt.com/reference/warehouse-profiles/fabricspark-profile"
"Documentation & usage" = "https://docs.getdbt.com/reference/resource-configs/fabricspark-configs"
"Changelog" = "https://github.com/microsoft/dbt-fabricspark/releases"
"Issue Tracker" = "https://github.com/microsoft/dbt-fabricspark/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/dbt/adapters/fabricspark/__version__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/dbt/adapters", "src/dbt/include"]
sources = ["src"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:.*datetime.datetime.utcnow.*:DeprecationWarning", # https://github.com/dbt-labs/dbt-core/issues/9791
]
env_files = ["test.env"]
testpaths = ["tests/unit", "tests/functional"]

[tool.ruff]
# Target Python versions
target-version = "py311"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I", "ERA"]
ignore = ["E203", "E501", "E741"]

[tool.ruff.lint.per-file-ignores]
# Ignore specific rules for certain files
"*/__init__.py" = ["F401"]
"tests/**/*" = ["ARG", "PLR0913"]

[tool.ruff.format]
# Formatting options
quote-style = "double"
indent-style = "space"
line-ending = "auto"

